// src/app/auth/models/SignUpData.ts
export class SignUpData {
    userId: number;
    userName: string;
    userPassword: string;
    email: string;
    // contactId: number;
    lastLoginDate: string;
    isActive: boolean;
    loginFailStreak: number;
    loginFailTotal: number;
    createdBy: number;
    dateCreated: string;
    modifiedBy: number;
    dateModified: string;
    deletedBy: number;
    dateDeleted: string;
    isDeleted: boolean;
    // primaryOfficeId: number;
    // profileId: number;
    dateLocked: string;
    managerId: number;
    userDisplayName: string
}